Fixed problems. Rewrited part of module. Stable version.
